Manage XML parsing per petition instead of parsing in constructor Extracted XMLHandlers to external classes
