+++ src/org/digitalcampus/oppia/utils/SearchUtils.java
@@ -29,6 +29,7 @@ import org.digitalcampus.oppia.model.Course;
 import org.digitalcampus.oppia.model.Lang;
 import org.digitalcampus.oppia.task.Payload;
 import org.digitalcampus.oppia.utils.storage.FileUtils;
+import org.digitalcampus.oppia.utils.xmlreaders.CourseXMLReader;
 
 import java.io.IOException;
 import java.util.ArrayList;
